+++ src/main/java/com/blog/domain/DataEntity.java
@@ -17,14 +17,14 @@ public class DataEntity {
 
     @Id
     @GeneratedValue
-    private Long id;
+    private Integer id;
 
-    public Long getId() {
+
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
-
 }
