+++ src/main/java/com/blog/repository/system/SysRoleRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SysRoleRepository extends JpaRepository<SysRole, Long> {
+public interface SysRoleRepository extends JpaRepository<SysRole, Integer> {
